==> Code/0.Preprocess.ipynb <==
# 预处理数据

包括：去除@，url，最小化

## 导入库

%time
import os
import sys
import warnings
if not sys.warnoptions:
    warnings.simplefilter("ignore")
    
import numpy as np
import pandas as pd
import sklearn

# Libraries and packages for text (pre-)processing 
import string
import re
import nltk

print("Python version:", sys.version)
print("Version info.:", sys.version_info)
print("pandas version:", pd.__version__)
print("numpy version:", np.__version__)
print("skearn version:", sklearn.__version__)
print("re version:", re.__version__)
print("nltk version:", nltk.__version__)
print("Operating system:", os.name)


## 对于文本数据进行清洗

df = pd.read_csv(r'C:\Software\Local Things (Coding)\comp\2025大学生建模比赛\代码\Twitter-Sentiment-Analysis-and-Tweet-Extraction-main\data\train.csv')
# df.head()
pred_df = pd.read_csv(r"C:\Software\Local Things (Coding)\comp\2025大学生建模比赛\数据\filtered_data.csv")
print(f"training data shape: {df.shape} \n testing data shape: {pred_df.shape}")

### 导入清洗数据函数

import re
def remove_URL(text):
    """
        Remove URLs from a sample string
    """
    return re.sub(r"https?://\S+|www\.\S+", "", text)

def remove_at_user(text):
    """
        Remove @user from a sample string
    """
    return re.sub(r'@[\w]+', '', text)

def remove_html(text):
    """
        Remove the html in sample text
    """
    html = re.compile(r"<.*?>|&([a-z0-9]+|#[0-9]{1,6}|#x[0-9a-f]{1,6});")
    return re.sub(html, "", text)

def remove_non_ascii(text):
    """
        Remove non-ASCII characters 
    """
    return re.sub(r'[^\x00-\x7f]',r'', text) # or ''.join([x for x in text if x in string.printable]) 

def remove_special_characters(text):
    """
        Remove special special characters, including symbols, emojis, and other graphic characters
    """
    emoji_pattern = re.compile(
        '['
        u'\U0001F600-\U0001F64F'  # emoticons
        u'\U0001F300-\U0001F5FF'  # symbols & pictographs
        u'\U0001F680-\U0001F6FF'  # transport & map symbols
        u'\U0001F1E0-\U0001F1FF'  # flags (iOS)
        u'\U00002702-\U000027B0'
        u'\U000024C2-\U0001F251'
        ']+',
        flags=re.UNICODE)
    return emoji_pattern.sub(r'', text)

def remove_punct(text):
    """
        Remove the punctuation
    """
#     return re.sub(r'[]!"$%&\'()*+,./:;=#@?[\\^_`{|}~-]+', "", text)
    return text.translate(str.maketrans('', '', string.punctuation))

def other_clean(text):
        """
            Other manual text cleaning techniques
        """
        # Typos, slang and other
        sample_typos_slang = {
                                "w/e": "whatever",
                                "usagov": "usa government",
                                "recentlu": "recently",
                                "ph0tos": "photos",
                                "amirite": "am i right",
                                "exp0sed": "exposed",
                                "<3": "love",
                                "luv": "love",
                                "amageddon": "armageddon",
                                "trfc": "traffic",
                                "16yr": "16 year"
                                }

        # Acronyms
        sample_acronyms =  { 
                            "mh370": "malaysia airlines flight 370",
                            "okwx": "oklahoma city weather",
                            "arwx": "arkansas weather",    
                            "gawx": "georgia weather",  
                            "scwx": "south carolina weather",  
                            "cawx": "california weather",
                            "tnwx": "tennessee weather",
                            "azwx": "arizona weather",  
                            "alwx": "alabama weather",
                            "usnwsgov": "united states national weather service",
                            "2mw": "tomorrow"
                            }

        
        # Some common abbreviations 
        sample_abbr = {
                        "$" : " dollar ",
                        "€" : " euro ",
                        "4ao" : "for adults only",
                        "a.m" : "before midday",
                        "a3" : "anytime anywhere anyplace",
                        "aamof" : "as a matter of fact",
                        "acct" : "account",
                        "adih" : "another day in hell",
                        "afaic" : "as far as i am concerned",
                        "afaict" : "as far as i can tell",
                        "afaik" : "as far as i know",
                        "afair" : "as far as i remember",
                        "afk" : "away from keyboard",
                        "app" : "application",
                        "approx" : "approximately",
                        "apps" : "applications",
                        "asap" : "as soon as possible",
                        "asl" : "age, sex, location",
                        "atk" : "at the keyboard",
                        "ave." : "avenue",
                        "aymm" : "are you my mother",
                        "ayor" : "at your own risk", 
                        "b&b" : "bed and breakfast",
                        "b+b" : "bed and breakfast",
                        "b.c" : "before christ",
                        "b2b" : "business to business",
                        "b2c" : "business to customer",
                        "b4" : "before",
                        "b4n" : "bye for now",
                        "b@u" : "back at you",
                        "bae" : "before anyone else",
                        "bak" : "back at keyboard",
                        "bbbg" : "bye bye be good",
                        "bbc" : "british broadcasting corporation",
                        "bbias" : "be back in a second",
                        "bbl" : "be back later",
                        "bbs" : "be back soon",
                        "be4" : "before",
                        "bfn" : "bye for now",
                        "blvd" : "boulevard",
                        "bout" : "about",
                        "brb" : "be right back",
                        "bros" : "brothers",
                        "brt" : "be right there",
                        "bsaaw" : "big smile and a wink",
                        "btw" : "by the way",
                        "bwl" : "bursting with laughter",
                        "c/o" : "care of",
                        "cet" : "central european time",
                        "cf" : "compare",
                        "cia" : "central intelligence agency",
                        "csl" : "can not stop laughing",
                        "cu" : "see you",
                        "cul8r" : "see you later",
                        "cv" : "curriculum vitae",
                        "cwot" : "complete waste of time",
                        "cya" : "see you",
                        "cyt" : "see you tomorrow",
                        "dae" : "does anyone else",
                        "dbmib" : "do not bother me i am busy",
                        "diy" : "do it yourself",
                        "dm" : "direct message",
                        "dwh" : "during work hours",
                        "e123" : "easy as one two three",
                        "eet" : "eastern european time",
                        "eg" : "example",
                        "embm" : "early morning business meeting",
                        "encl" : "enclosed",
                        "encl." : "enclosed",
                        "etc" : "and so on",
                        "faq" : "frequently asked questions",
                        "fawc" : "for anyone who cares",
                        "fb" : "facebook",
                        "fc" : "fingers crossed",
                        "fig" : "figure",
                        "fimh" : "forever in my heart", 
                        "ft." : "feet",
                        "ft" : "featuring",
                        "ftl" : "for the loss",
                        "ftw" : "for the win",
                        "fwiw" : "for what it is worth",
                        "fyi" : "for your information",
                        "g9" : "genius",
                        "gahoy" : "get a hold of yourself",
                        "gal" : "get a life",
                        "gcse" : "general certificate of secondary education",
                        "gfn" : "gone for now",
                        "gg" : "good game",
                        "gl" : "good luck",
                        "glhf" : "good luck have fun",
                        "gmt" : "greenwich mean time",
                        "gmta" : "great minds think alike",
                        "gn" : "good night",
                        "g.o.a.t" : "greatest of all time",
                        "goat" : "greatest of all time",
                        "goi" : "get over it",
                        "gps" : "global positioning system",
                        "gr8" : "great",
                        "gratz" : "congratulations",
                        "gyal" : "girl",
                        "h&c" : "hot and cold",
                        "hp" : "horsepower",
                        "hr" : "hour",
                        "hrh" : "his royal highness",
                        "ht" : "height",
                        "ibrb" : "i will be right back",
                        "ic" : "i see",
                        "icq" : "i seek you",
                        "icymi" : "in case you missed it",
                        "idc" : "i do not care",
                        "idgadf" : "i do not give a damn fuck",
                        "idgaf" : "i do not give a fuck",
                        "idk" : "i do not know",
                        "ie" : "that is",
                        "i.e" : "that is",
                        "ifyp" : "i feel your pain",
                        "IG" : "instagram",
                        "iirc" : "if i remember correctly",
                        "ilu" : "i love you",
                        "ily" : "i love you",
                        "imho" : "in my humble opinion",
                        "imo" : "in my opinion",
                        "imu" : "i miss you",
                        "iow" : "in other words",
                        "irl" : "in real life",
                        "j4f" : "just for fun",
                        "jic" : "just in case",
                        "jk" : "just kidding",
                        "jsyk" : "just so you know",
                        "l8r" : "later",
                        "lb" : "pound",
                        "lbs" : "pounds",
                        "ldr" : "long distance relationship",
                        "lmao" : "laugh my ass off",
                        "lmfao" : "laugh my fucking ass off",
                        "lol" : "laughing out loud",
                        "ltd" : "limited",
                        "ltns" : "long time no see",
                        "m8" : "mate",
                        "mf" : "motherfucker",
                        "mfs" : "motherfuckers",
                        "mfw" : "my face when",
                        "mofo" : "motherfucker",
                        "mph" : "miles per hour",
                        "mr" : "mister",
                        "mrw" : "my reaction when",
                        "ms" : "miss",
                        "mte" : "my thoughts exactly",
                        "nagi" : "not a good idea",
                        "nbc" : "national broadcasting company",
                        "nbd" : "not big deal",
                        "nfs" : "not for sale",
                        "ngl" : "not going to lie",
                        "nhs" : "national health service",
                        "nrn" : "no reply necessary",
                        "nsfl" : "not safe for life",
                        "nsfw" : "not safe for work",
                        "nth" : "nice to have",
                        "nvr" : "never",
                        "nyc" : "new york city",
                        "oc" : "original content",
                        "og" : "original",
                        "ohp" : "overhead projector",
                        "oic" : "oh i see",
                        "omdb" : "over my dead body",
                        "omg" : "oh my god",
                        "omw" : "on my way",
                        "p.a" : "per annum",
                        "p.m" : "after midday",
                        "pm" : "prime minister",
                        "poc" : "people of color",
                        "pov" : "point of view",
                        "pp" : "pages",
                        "ppl" : "people",
                        "prw" : "parents are watching",
                        "ps" : "postscript",
                        "pt" : "point",
                        "ptb" : "please text back",
                        "pto" : "please turn over",
                        "qpsa" : "what happens", #"que pasa",
                        "ratchet" : "rude",
                        "rbtl" : "read between the lines",
                        "rlrt" : "real life retweet", 
                        "rofl" : "rolling on the floor laughing",
                        "roflol" : "rolling on the floor laughing out loud",
                        "rotflmao" : "rolling on the floor laughing my ass off",
                        "rt" : "retweet",
                        "ruok" : "are you ok",
                        "sfw" : "safe for work",
                        "sk8" : "skate",
                        "smh" : "shake my head",
                        "sq" : "square",
                        "srsly" : "seriously", 
                        "ssdd" : "same stuff different day",
                        "tbh" : "to be honest",
                        "tbs" : "tablespooful",
                        "tbsp" : "tablespooful",
                        "tfw" : "that feeling when",
                        "thks" : "thank you",
                        "tho" : "though",
                        "thx" : "thank you",
                        "tia" : "thanks in advance",
                        "til" : "today i learned",
                        "tl;dr" : "too long i did not read",
                        "tldr" : "too long i did not read",
                        "tmb" : "tweet me back",
                        "tntl" : "trying not to laugh",
                        "ttyl" : "talk to you later",
                        "u" : "you",
                        "u2" : "you too",
                        "u4e" : "yours for ever",
                        "utc" : "coordinated universal time",
                        "w/" : "with",
                        "w/o" : "without",
                        "w8" : "wait",
                        "wassup" : "what is up",
                        "wb" : "welcome back",
                        "wtf" : "what the fuck",
                        "wtg" : "way to go",
                        "wtpa" : "where the party at",
                        "wuf" : "where are you from",
                        "wuzup" : "what is up",
                        "wywh" : "wish you were here",
                        "yd" : "yard",
                        "ygtr" : "you got that right",
                        "ynk" : "you never know",
                        "zzz" : "sleeping bored and tired"
                        }
            
        sample_typos_slang_pattern = re.compile(r'(?<!\w)(' + '|'.join(re.escape(key) for key in sample_typos_slang.keys()) + r')(?!\w)')
        sample_acronyms_pattern = re.compile(r'(?<!\w)(' + '|'.join(re.escape(key) for key in sample_acronyms.keys()) + r')(?!\w)')
        sample_abbr_pattern = re.compile(r'(?<!\w)(' + '|'.join(re.escape(key) for key in sample_abbr.keys()) + r')(?!\w)')
        
        text = sample_typos_slang_pattern.sub(lambda x: sample_typos_slang[x.group()], text)
        text = sample_acronyms_pattern.sub(lambda x: sample_acronyms[x.group()], text)
        text = sample_abbr_pattern.sub(lambda x: sample_abbr[x.group()], text)
        
        return text

### 清理文本数据

import contractions
from textblob import TextBlob
from nltk.tokenize import RegexpTokenizer
from autocorrect import Speller

def preprocess(df, cols='text'):
    # 处理数据
    cols_after = cols + '_after'
    df[cols] = df[cols].astype(str)  # 转换为字符串
    print(f'Original...{df.shape}')
    print('============================')
    
    print(f'Preprocessing...{df.shape}')
    print(f'Lowercase...{df.shape}')
    df[cols_after] = df[cols].apply(lambda x: x.lower())  # 全部转为小写
    print(f'Contraction...{df.shape}')
    df[cols_after] = df[cols_after].apply(lambda x: contractions.fix(x))  # 处理缩写
    print(f'Remove URL...{df.shape}')
    df[cols_after] = df[cols_after].apply(lambda x: remove_URL(x))  # 去除URL
    print(f'Remove HTML...{df.shape}')
    df[cols_after] = df[cols_after].apply(lambda x: remove_html(x))  # 去除HTML标签
    print(f'Remove @user...{df.shape}')
    df[cols_after] = df[cols_after].apply(lambda x: remove_at_user(x))  # 去除@用户
    print(f'Remove non-ascii...{df.shape}')
    df[cols_after] = df[cols_after].apply(lambda x: remove_non_ascii(x))  # 去除表情符号
    print(f'Remove special characters...{df.shape}')
    df[cols_after] = df[cols_after].apply(lambda x: remove_special_characters(x))  # 去除特殊字符
    print(f'Remove punctuations...{df.shape}')
    df[cols_after] = df[cols_after].apply(lambda x: remove_punct(x))  # 去除标点符号
    print(f'Other clean...{df.shape}')
    df[cols_after] = df[cols_after].apply(lambda x: other_clean(x))  # 其他清理
    # print()
    # print(f'Auto correct...')
    # df[cols_after] = df[cols_after].apply(lambda x: TextBlob(x).correct().string)  # 自动纠错

    # 分词和拼写检查
    tokenizer = RegexpTokenizer(r'\w+')
    spell = Speller(lang="en")
    print(f'Tokenization...{df.shape}')
    df[f"{cols}_token"] = df[cols_after].apply(lambda x: tokenizer.tokenize(x))  # 去除标点符号并分词
    print(f'Spell check...{df.shape}')
    df[cols_after] = df[cols_after].apply(lambda x: [spell(i) for i in x])  # 拼写检查
    print(f'Done...{df.shape}')
    # df[cols_after] = df[cols_after].apply(lambda x: re.sub(r'^\s*|\s\s*', ' ', x).strip()) # 空格分词
    # print(f'Space tokenization...')
    return df


import os
os.chdir(r"C:\Software\Local Things (Coding)\comp\2025大学生建模比赛\代码\比赛项目代码\处理后数据集")

df_processed = preprocess(df,cols = 'text')
df_processed.to_csv('train_cleaned.csv',index=False)
# df_processed.head()
pred_df_processed = preprocess(pred_df,cols = 'full_text')
pred_df_processed.to_csv('test_cleaned.csv',index=False)

# 词云图

## 训练数据

import pandas as pd
df = pd.read_csv(r"C:\Software\Local Things (Coding)\comp\2025大学生建模比赛\代码\比赛项目代码\处理后数据集\train_cleaned.csv")
print(df.columns)

from wordcloud import WordCloud
import matplotlib.pyplot as plt
import pandas as pd

# 定义一个函数来生成词云图
def generate_wordcloud(text, title):
    wordcloud = WordCloud(width=800, height=400, background_color='white').generate(text)
    plt.imshow(wordcloud, interpolation='bilinear')
    plt.title(title)
    plt.axis('off')

# 提取每个类别中的 tokens 并合并成字符串
categories = ['positive', 'negative', 'neutral']
texts = {}

for category in categories:
    category_tokens = []
    for tokens in df[df['sentiment'] == category]['text_token']:
        category_tokens.extend(eval(tokens))  # 使用 eval() 转换字符串表示的列表
    texts[category] = ' '.join(category_tokens)

# 创建子图并显示每个类别的词云图
plt.figure(figsize=(15, 5))

for i, category in enumerate(categories):
    plt.subplot(1, 3, i + 1)  # 1 行 3 列的子图布局
    generate_wordcloud(texts[category], category.capitalize())

plt.tight_layout()  # 调整子图之间的间距
plt.savefig('train data wordcloud.png')
plt.show()


import pandas as pd
import matplotlib.pyplot as plt
from collections import Counter


plt.rcParams['font.sans-serif'] = ['SimHei']  # 使用黑体
plt.rcParams['axes.unicode_minus'] = False  # 解决负号显示问题

# 定义一个函数来计算每个token的概率
def calculate_token_probabilities(tokens):
    token_counts = Counter(tokens)
    total_tokens = len(tokens)
    token_probabilities = {token: count / total_tokens for token, count in token_counts.items()}
    return token_probabilities

# 提取每个类别中的 tokens 并合并成列表
categories = ['positive', 'negative', 'neutral']
category_tokens = {category: [] for category in categories}

for category in categories:
    for tokens in df[df['sentiment'] == category]['text_token']:
        category_tokens[category].extend(eval(tokens))  # 使用 eval() 转换字符串表示的列表

# 计算每个类别中每个token的概率
token_probabilities = {category: calculate_token_probabilities(tokens) for category, tokens in category_tokens.items()}

# 为了绘制概率分布图，我们需要将token和其概率转换为有序的列表
sorted_token_probabilities = {category: sorted(probs.items(), key=lambda item: item[1], reverse=True) for category, probs in token_probabilities.items()}

# 创建子图并显示每个类别的概率分布图
plt.figure(figsize=(15, 5))

for i, category in enumerate(categories):
    plt.subplot(1, 3, i + 1)  # 1 行 3 列的子图布局
    tokens, probabilities = zip(*sorted_token_probabilities[category][:20])  # 只取前20个token
    plt.bar(tokens, probabilities)
    plt.title(f'{category.capitalize()} Token 概率分布图')
    plt.xlabel('Tokens')
    plt.ylabel('概率')
    plt.xticks(rotation=45)  # 旋转x轴标签以便更好地显示

plt.tight_layout()  # 调整子图之间的间距
plt.savefig('train_data_token_distribution.png')
plt.show()


## 预测数据

import pandas as pd
df = pd.read_csv(r"C:\Software\Local Things (Coding)\comp\2025大学生建模比赛\代码\比赛项目代码\处理后数据集\test_cleaned.csv")
print(df.columns)
# print(df.head())

from wordcloud import WordCloud
import matplotlib.pyplot as plt

# 提取 full_text_token 列，并将所有的 tokens 合并成一个大的文本字符串
all_tokens = []
for tokens in df['full_text_token']:
    all_tokens.extend(eval(tokens))  # 使用 eval() 转换字符串表示的列表

# 将所有 tokens 合并成一个大字符串
text = ' '.join(all_tokens)

# 生成词云图
wordcloud = WordCloud(width=800, height=400, background_color='white').generate(text)

# 显示词云图
plt.figure(figsize=(10, 5))
plt.imshow(wordcloud, interpolation='bilinear')
plt.axis('off')  # 不显示坐标轴
plt.show()

import pandas as pd
from wordcloud import WordCloud
import matplotlib.pyplot as plt

# 加载数据
data_path = r"C:\Software\Local Things (Coding)\comp\2025大学生建模比赛\代码\比赛项目代码\处理后数据集\test_cleaned.csv"
df = pd.read_csv(data_path)

# 生成整体词云
all_tokens = []
for tokens in df['full_text_token']:
    all_tokens.extend(eval(tokens))
    
overall_text = ' '.join(all_tokens)
overall_wordcloud = WordCloud(width=800, height=400, background_color='white').generate(overall_text)

# 准备子图布局
unique_names = df['name'].unique()
num_total = len(unique_names) + 1  # 增加整体词云
cols = 4
rows = (num_total // cols) + (1 if num_total % cols != 0 else 0)

# 创建包含整体词云的画布
fig, axes = plt.subplots(nrows=rows, ncols=cols, figsize=(15, rows * 2))
axes = axes.flatten()

# 第1个子图绘制整体词云
axes[0].imshow(overall_wordcloud, interpolation='bilinear')
axes[0].axis('off')
axes[0].set_title("Overall WordCloud", fontsize=12)

# 遍历每个name绘制子图
for idx, (name, group) in enumerate(df.groupby('name'), start=1):  # 从索引1开始
    name_tokens = []
    for tokens in group['full_text_token']:
        name_tokens.extend(eval(tokens))
    
    text = ' '.join(name_tokens)
    wordcloud = WordCloud(width=800, height=400, background_color='white').generate(text)
    
    axes[idx].imshow(wordcloud, interpolation='bilinear')
    axes[idx].axis('off')
    axes[idx].set_title(f"{name}", fontsize=12)

# 隐藏空白子图
for idx in range(num_total, len(axes)):
    fig.delaxes(axes[idx])

# 调整布局并保存
plt.tight_layout()
plt.savefig('词云图.png', dpi=300, bbox_inches='tight')
plt.show()


from PIL import Image
import os

# 设置头像文件夹路径
folder_path = r'C:\Users\Fisher Man\Downloads\TweetCrawler\头像列表'  # 替换为你的文件夹路径

# 获取文件夹中所有文件的路径
image_files = [f for f in os.listdir(folder_path) if f.endswith(('.jpg', '.png'))]

# 创建完整的文件路径列表
image_paths = [os.path.join(folder_path, file) for file in image_files]

# 加载所有头像图片
images = [Image.open(img_path) for img_path in image_paths]

# 调整图片的尺寸，以便统一拼接
image_size = (100, 100)  # 统一尺寸 100x100，可以根据需要调整
images_resized = [img.resize(image_size) for img in images]

# 拼接设置（比如按每行4个头像）
cols = 6
rows = (len(images_resized) // cols) + (1 if len(images_resized) % cols != 0 else 0)

# 创建一个新图像，用于拼接所有头像
# 计算拼接后的图片大小
width = cols * image_size[0]
height = rows * image_size[1]
result_image = Image.new('RGB', (width, height))

# 将头像按行列排列
for index, img in enumerate(images_resized):
    row = index // cols
    col = index % cols
    # 计算每个头像在大图中的位置
    result_image.paste(img, (col * image_size[0], row * image_size[1]))

# 保存结果图像
result_image.save(r'拼接后的头像图.png')

# 展示拼接后的图像
result_image.show()
